# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
...
        "/api/arquivo-musica": {
            "post": {
                "tags": [
                    "Arquivo Musica MP3"
                ],
                "summary": "Faz o upload de um arquivo de musica",
                "description": "Upload de Arquivo",
                "operationId": "7d16d66fd7a7ad79c71a89450426306c",
                "parameters": [
                    {
                        "name": "album_id",
                        "in": "query",
                        "description": "Nº de identificação do Album",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Arquivo de música enviado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "file": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro no upload do MP3"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/artistaalbum": {
            "post": {
                "tags": [
                    "Artistas Albuns"
                ],
                "summary": "Cadastra vínculo de Artista com álbum",
                "operationId": "a23623e6b037ed5e790505568ed30b73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "art_id",
                                    "alb_id"
                                ],
                                "properties": {
                                    "art_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "alb_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vínculo criado"
                    },
                    "409": {
                        "description": "Vínculo já existente"
                    },
                    "404": {
                        "description": "Artista ou Álbum não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/artistaalbum/{id}": {
            "delete": {
                "tags": [
                    "Artistas Albuns"
                ],
                "summary": "Remove um vínculo",
                "operationId": "90b066a2cd7826e2b1f3fa6438f90c3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/musica": {
            "get": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Todas as Musicas",
                "description": "Lista todas as Musicas",
                "operationId": "01de9509d3e6ccaa97ce184616faba9b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº da página",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Musicas"
                ],
                "summary": "Cadastra nova Musica",
                "description": "Registra uma nova Musica",
                "operationId": "c703a59db1ae0961c25ba18b8721f70c",
                "parameters": [
                    {
                        "name": "mus_titulo",
                        "in": "query",
                        "description": "Titulo da Musica",
                        "required": true
                    },
                    {
                        "name": "album_id",
                        "in": "query",
                        "description": "Nº de identificação do Album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Musica não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}